# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Moviestore API",
        "description": "This is the documentation for the moviestore API.",
        "contact": {
            "email": "setty.095@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/",
            "description": "Moviestore API"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Logs user into the system",
                "description": "",
                "operationId": "loginUser",
                "requestBody": {
                    "description": "Movie object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "$ref": "#/components/schemas/User/properties/email"
                                    },
                                    "password": {
                                        "$ref": "#/components/schemas/User/properties/password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "405": {
                        "description": "Method Not Allowed"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Logs out current logged in user session",
                "description": "",
                "operationId": "logoutUser",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "apiAuth": [
                            "write:movies",
                            "read:movies"
                        ]
                    }
                ]
            }
        },
        "/movies": {
            "get": {
                "tags": [
                    "Movie"
                ],
                "summary": "Get movies",
                "description": "Returns list of movies",
                "operationId": "getMovies",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "405": {
                        "description": "Method Not Allowed"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/movies/{movieId}": {
            "get": {
                "tags": [
                    "Movie"
                ],
                "summary": "Find movie by ID",
                "description": "Returns a single movie",
                "operationId": "getMovieById",
                "parameters": [
                    {
                        "name": "movieId",
                        "in": "path",
                        "description": "ID of movie to return",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Movie not found"
                    },
                    "405": {
                        "description": "Method Not Allowed"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/search-movies": {
            "get": {
                "tags": [
                    "Movie"
                ],
                "summary": "Search for movies.",
                "description": "Returns list of related movies",
                "operationId": "searchMovies",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "405": {
                        "description": "Method Not Allowed"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "c85c9c7c-dbb8-4d62-b1a9-f5585ee208d6"
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "johndoe@gmail.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "password"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            }
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token bearer Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}